Update to 147.7715.
Update to commit 3c790322ed50cac026372afdb79e110d8eb101ea

+++ components/services/storage/dom_storage/local_storage_impl.cc
@@ -13,19 +13,12 @@
 #include <string_view>
 #include <utility>
 
-#include "base/barrier_closure.h"
 #include "base/byte_size.h"
-#include "base/debug/dump_without_crashing.h"
-#include "base/files/file_enumerator.h"
 #include "base/files/file_path.h"
 #include "base/functional/bind.h"
-#include "base/functional/callback_helpers.h"
 #include "base/memory/raw_ptr.h"
 #include "base/metrics/histogram_functions.h"
 #include "base/metrics/histogram_macros.h"
-#include "base/numerics/safe_conversions.h"
-#include "base/strings/string_number_conversions.h"
-#include "base/strings/string_view_util.h"
 #include "base/strings/stringprintf.h"
 #include "base/system/sys_info.h"
 #include "base/task/thread_pool.h"
@@ -35,14 +28,10 @@
 #include "components/services/storage/dom_storage/db_status.h"
 #include "components/services/storage/dom_storage/dom_storage_constants.h"
 #include "components/services/storage/dom_storage/dom_storage_database.h"
-#include "components/services/storage/dom_storage/leveldb/local_storage_leveldb.h"
 #include "components/services/storage/dom_storage/leveldb_status_helper.h"
 #include "components/services/storage/dom_storage/storage_area_impl.h"
-#include "components/services/storage/public/cpp/constants.h"
-#include "mojo/public/cpp/bindings/self_owned_receiver.h"
 #include "storage/common/database/database_identifier.h"
 #include "third_party/blink/public/common/storage_key/storage_key.h"
-#include "url/gurl.h"
 
 namespace storage {
 namespace {
@@ -101,8 +90,7 @@ class LocalStorageImpl::StorageAreaHolder final
         storage_key_(storage_key),
         area_(context_->database_.get(),
               base::MakeRefCounted<DomStorageDatabase::SharedMapLocator>(
-                  DomStorageDatabase::MapLocator(kLocalStorageSessionId,
-                                                 storage_key_)),
+                  DomStorageDatabase::MapLocator(storage_key_)),
               this,
               createOptions()) {}
 
@@ -115,6 +103,12 @@ class LocalStorageImpl::StorageAreaHolder final
     if (storage_area()->empty()) {
       return;
     }
+
+    if (!context_->database_) {
+      // The database does not exist.
+      return;
+    }
+
     // We should not write last_accessed if the data will be purged.
     if (context_->origins_to_purge_on_shutdown_.find(storage_key_.origin()) !=
             context_->origins_to_purge_on_shutdown_.end() ||
@@ -124,10 +118,10 @@ class LocalStorageImpl::StorageAreaHolder final
       return;
     }
 
-    // Update the storage area map's last access time.
+    // Update the storage area map's last access time in the database.
     DomStorageDatabase::Metadata usage;
     usage.map_metadata.push_back({
-        .map_locator{kLocalStorageSessionId, storage_key_},
+        .map_locator{storage_key_},
         .last_accessed{base::Time::Now()},
     });
     context_->database_->PutMetadata(
@@ -246,10 +240,10 @@ void LocalStorageImpl::DeleteStorage(const blink::StorageKey& storage_key,
     found->second->storage_area()->ScheduleImmediateCommit();
   } else if (database_) {
     std::vector<DomStorageDatabase::MapLocator> maps_to_delete;
-    maps_to_delete.emplace_back(kLocalStorageSessionId, storage_key);
+    maps_to_delete.emplace_back(storage_key);
 
     database_->DeleteStorageKeysFromSession(
-        kLocalStorageSessionId, /*metadata_to_delete=*/{storage_key},
+        /*session_id=*/std::string(), /*metadata_to_delete=*/{storage_key},
         std::move(maps_to_delete),
         base::BindOnce([](base::OnceClosure callback,
                           DbStatus) { std::move(callback).Run(); },
@@ -321,7 +315,8 @@ void LocalStorageImpl::ShutDown() {
       area->CancelAllPendingRequests();
     }
 
-    if (!force_keep_session_state_ && !origins_to_purge_on_shutdown_.empty()) {
+    if (database_ && !force_keep_session_state_ &&
+        !origins_to_purge_on_shutdown_.empty()) {
       database_->PurgeOriginsForShutdown(
           std::move(origins_to_purge_on_shutdown_));
     }
@@ -469,7 +464,7 @@ void LocalStorageImpl::PurgeAllStorageAreas() {
 }
 
 void LocalStorageImpl::InitiateConnection(bool in_memory_only) {
-  DCHECK_EQ(connection_state_, CONNECTION_IN_PROGRESS);
+  CHECK_EQ(connection_state_, CONNECTION_IN_PROGRESS);
 
   if (!storage_partition_directory_.empty() &&
       storage_partition_directory_.IsAbsolute() && !in_memory_only) {
@@ -492,7 +487,6 @@ void LocalStorageImpl::InitiateConnection(bool in_memory_only) {
 }
 
 void LocalStorageImpl::OnDatabaseOpened(DbStatus status) {
-  status.Log("Storage.LocalStorage.OpenDatabase", in_memory_);
   if (!status.ok()) {
     // If we failed to open the database, try to delete and recreate the
     // database, or ultimately fallback to an in-memory database.
@@ -504,7 +498,7 @@ void LocalStorageImpl::OnDatabaseOpened(DbStatus status) {
 }
 
 void LocalStorageImpl::OnConnectionFinished() {
-  DCHECK_EQ(connection_state_, CONNECTION_IN_PROGRESS);
+  CHECK_EQ(connection_state_, CONNECTION_IN_PROGRESS);
   // If connection was opened successfully, reset tried_to_recreate_during_open_
   // to enable recreating the database on future errors.
   if (database_)
@@ -575,7 +569,7 @@ void LocalStorageImpl::OnDBDestroyed(bool recreate_in_memory, DbStatus status) {
 
 LocalStorageImpl::StorageAreaHolder* LocalStorageImpl::GetOrCreateStorageArea(
     const blink::StorageKey& storage_key) {
-  DCHECK_EQ(connection_state_, CONNECTION_FINISHED);
+  CHECK_EQ(connection_state_, CONNECTION_FINISHED);
   auto found = areas_.find(storage_key);
   if (found != areas_.end()) {
     return found->second.get();
@@ -590,6 +584,8 @@ LocalStorageImpl::StorageAreaHolder* LocalStorageImpl::GetOrCreateStorageArea(
 }
 
 void LocalStorageImpl::RetrieveStorageUsage(GetUsageCallback callback) {
+  CHECK_EQ(connection_state_, ConnectionState::CONNECTION_FINISHED);
+
   if (!database_) {
     // If for whatever reason no database is available, no storage is
     // used, so return an array only containing the current areas.
@@ -657,9 +653,15 @@ void LocalStorageImpl::GetStatistics(size_t* total_cache_size,
 }
 
 void LocalStorageImpl::OnCommitResult(DbStatus status) {
-  DCHECK(connection_state_ == CONNECTION_FINISHED) << connection_state_;
   if (status.ok()) {
     commit_error_count_ = 0;
+    return;
+  }
+
+  if (connection_state_ != CONNECTION_FINISHED) {
+    // Previous commit errors deleted and recreated the database below.  Ignore
+    // additional errors from the old database while waiting for the new
+    // database to open.
     return;
   }
 
@@ -681,7 +683,7 @@ void LocalStorageImpl::OnCommitResult(DbStatus status) {
 }
 
 void LocalStorageImpl::DeleteStaleStorageAreas() {
-  if (!database_) {
+  if (!database_ || connection_state_ != CONNECTION_FINISHED) {
     // Due to the delay before LocalStorageImpl::DeleteStaleStorageAreas is invoked
     // it's possible `database_` existed before, but no longer.
     return;
@@ -693,7 +695,7 @@ void LocalStorageImpl::DeleteStaleStorageAreas() {
 
 void LocalStorageImpl::OnGotMetaDataToDeleteStaleStorageAreas(
     StatusOr<DomStorageDatabase::Metadata> all_metadata) {
-  if (!database_) {
+  if (!database_ || connection_state_ != CONNECTION_FINISHED) {
     // This method is provided as a callback to an off thread task. Between the
     // time that the task is posted and now when this callback is invoked, the
     // `database_` member may have been reset.
@@ -732,7 +734,7 @@ void LocalStorageImpl::OnGotMetaDataToDeleteStaleStorageAreas(
       // If the storage area has not been accessed or modified within 400 days
       // it can be cleared.
       stale_storage_keys.push_back(storage_key);
-      maps_to_delete.emplace_back(kLocalStorageSessionId, storage_key);
+      maps_to_delete.emplace_back(storage_key);
     } else if ((storage_key.nonce().has_value() ||
                 storage_key.top_level_site().opaque()) &&
                (base::Time::Now() - accessed_or_modified_time) >=
@@ -740,7 +742,7 @@ void LocalStorageImpl::OnGotMetaDataToDeleteStaleStorageAreas(
       // If the storage area has not been accessed or modified in this browsing
       // session and is transient (has a nonce) then it can be cleared.
       stale_storage_keys.push_back(storage_key);
-      maps_to_delete.emplace_back(kLocalStorageSessionId, storage_key);
+      maps_to_delete.emplace_back(storage_key);
       orphans_found++;
     }
   }
@@ -751,7 +753,7 @@ void LocalStorageImpl::OnGotMetaDataToDeleteStaleStorageAreas(
   // Delete stale storage areas and count results.
   size_t deleted_count = stale_storage_keys.size();
   database_->DeleteStorageKeysFromSession(
-      kLocalStorageSessionId,
+      /*session_id=*/std::string(),
       /*metadata_to_delete=*/std::move(stale_storage_keys),
       std::move(maps_to_delete),
       base::BindOnce(
